requirements.txt
fastapi
uvicorn
python-dotenv
groq
python-multipart
langchain-text-splitters
requests
